cookbook_webservice/requirements.txt
django
djangorestframework
django-cors-headers
Pillow
django-rest-auth
djangorestframework-jwt
channels
channels_redis